63 관리자 고객관리 응답 데이터 수정

+++ src/main/java/com/meongnyang/shop/dto/response/admin/RespProductDetailDto.java
@@ -19,7 +19,7 @@ public class RespProductDetailDto {
     private PetGroup petGroup;
     private Category category;
     private Long productPrice;
-    private int productPriceDiscount;
+    private Long productPriceDiscount;
     private String productDetail;
     private String productBrand;
     private String productModel;
